Fixed signed/unsigned compilation warning by making wxAuiManager::GetActivePane() return value type consistent with other GetActive...() API that can return wxNOT_FOUND.

+++ include/wx/aui/framemanager.h
@@ -702,7 +702,7 @@ public:
     void RestoreMaximizedPane();
 
     int SetActivePane(wxWindow* activePane);
-    size_t GetActivePane(wxWindow* focus) const;
+    int GetActivePane(wxWindow* focus) const;
 
     virtual wxAuiFloatingFrame* CreateFloatingFrame(wxWindow* parent, const wxAuiPaneInfo& p);
     virtual bool CanDockPanel(const wxAuiPaneInfo& p);
